pyproject.toml
[project]
name = "rainforcement-learning"
version = "0.1.0"
description = ""
authors = [
    {name = "kmykprn",email = "kmykprn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "torch (==2.5.1)",
    "numpy (>=2.2.4,<3.0.0)",
    "beartype (>=0.20.2,<0.21.0)",
    "jaxtyping (>=0.3.1,<0.4.0)",
    "torchvision (>=0.20.1,<0.21.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch_cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "torch_cu121"}
torchvision = {source = "torch_cu121"}
